+++ TabletopTweaks-Base/NewContent/Feats/MetamagicFeats/PiercingSpell.cs
@@ -102,7 +102,10 @@ namespace TabletopTweaks.Base.NewContent.Feats.MetamagicFeats {
         public static void UpdateSpells() {
             if (TTTContext.AddedContent.Feats.IsDisabled("MetamagicPiercingSpell")) { return; }
 
-            var spells = SpellTools.GetAllSpells();
+            var spells = SpellTools.GetAllSpells()
+                .SelectMany(s => s.AbilityAndVariants())
+                .SelectMany(s => s.AbilityAndStickyTouch())
+                .ToArray();
             foreach (var spell in spells) {
                 bool validPiercing = spell.AbilityAndVariants()
                     .SelectMany(s => s.AbilityAndStickyTouch())