+++ TabletopTweaks-Base/NewContent/Feats/MetamagicFeats/RimeSpell.cs
@@ -129,7 +129,10 @@ namespace TabletopTweaks.Base.NewContent.Feats.MetamagicFeats {
         public static void UpdateSpells() {
             if (TTTContext.AddedContent.Feats.IsDisabled("MetamagicRimeSpell")) { return; }
 
-            var spells = SpellTools.GetAllSpells();
+            var spells = SpellTools.GetAllSpells()
+                .SelectMany(s => s.AbilityAndVariants())
+                .SelectMany(s => s.AbilityAndStickyTouch())
+                .ToArray();
             foreach (var spell in spells) {
                 bool isColdSpell = spell.AbilityAndVariants()
                     .SelectMany(s => s.AbilityAndStickyTouch())